Add animation to lab7.js
The following code uses arrays to move the butterflies. There is a variable to edit how many butterflies (or "movers") there are in the Canvas. Currently, they all follow the path of a parametric equation. However, there is a section of code in the setup that can customize the center points of the the center points of the parametric equation and the max distance from the center of the parametric equation too.